rename SetValueContent to valuemodal
