+++ frontend/src/modules/controllers/ValueModal.tsx
@@ -1,14 +1,19 @@
 /**
- * @deprecated
- * @summary Re-usable wrapper components to display & control value
+ * ValueModal
+ *
+ * @component A container for setting the value of modal settings in multistep popup.
+ * This is a modified version of ValueModal
+ *
+ * Uses the [[ContentProps]] interface
+ * @returns JSX.Element
  */
-import React, { useCallback, useEffect } from 'react';
-import { makeStyles, Theme, Grid, Button, Typography } from '@material-ui/core';
-import { shallowEqual, useSelector } from 'react-redux';
-import ValueClicker from './ValueClicker';
-import ModalPopup from './ModalPopup';
+
+import { Grid, makeStyles, Theme, Typography } from '@material-ui/core';
+import React, { useRef, useEffect, useCallback } from 'react';
+import { useSelector, shallowEqual } from 'react-redux';
 import { getRotaryEncoder } from '../../store/controller/selectors';
-import { RotaryEncodeController } from './RotaryEncodeController';
+import RotaryEncodeController from './RotaryEncodeController';
+import ValueClicker from './ValueClicker';
 
 const useStyles = makeStyles((theme: Theme) => ({
   contentContainer: {
@@ -36,27 +41,13 @@ const useStyles = makeStyles((theme: Theme) => ({
 }));
 
 /**
- * @deprecated
- * @typedef SettingAdjustProps
+ * @typedef ContentProps
  *
- * Interface for the adjusting props settings
+ * Props interface for the setting value
  *
- * @prop {number} committedSetting updated value
- */
-export interface SettingAdjustProps {
-  committedSetting: number;
-}
-
-/**
- * @deprecated
- * @typedef Props
- *
- * Props interface for the value model
- *
- * @prop {string} label desc for label
- * @prop {string} units desc for units
+ * @prop {string} label Label for the value
+ * @prop {string} units Unit measurement of the value
  * @prop {number} committedSetting current value
- * TODO: disableSetNewButton can be removed as it's unused.
  * @prop {boolean} disableSetNewButton Configuration to show/hide `Set New` outside modal
  * @prop {function} requestCommitSetting Callback when value changes
  * @prop {function} updateModalStatus Callback when modal status changes
@@ -65,11 +56,10 @@ export interface SettingAdjustProps {
  * @prop {number} max Maximum above which value cannot increment
  *
  */
-interface Props {
+interface ContentProps {
   label: string;
   units?: string;
   committedSetting: number;
-  disableSetNewButton?: boolean;
   requestCommitSetting(setting: number): void;
   updateModalStatus?(status: boolean): void;
   openModal?: boolean;
@@ -77,49 +67,51 @@ interface Props {
   max?: number;
 }
 
-/**
- * @deprecated
- * ValueModal
- *
- * @component A container for displaying value modal settings.
- *
- * Uses the [[Props]] interface
- *
- * @returns JSX.Element
- */
-export const ValueModal = ({
+const ValueModal = ({
   label,
   units,
   committedSetting,
-  disableSetNewButton = false,
   openModal = false,
   updateModalStatus,
   requestCommitSetting,
   min = 0,
   max = 100,
-}: Props): JSX.Element => {
+}: ContentProps): JSX.Element => {
   const classes = useStyles();
+  const isInitialMount = useRef(true);
   const rotaryEncoder = useSelector(getRotaryEncoder, shallowEqual);
   /**
    * State to manage modal status
    */
-  const [open, setOpen] = React.useState(false);
+  const [open, setOpen] = React.useState(openModal);
   /**
-   * State to manage value
+   * Initalization to set default value & modal status
    */
   const [value, setValue] = React.useState(committedSetting);
+  /**
+   * State to manage rotary encoder acitvity
+   */
+  const [isRotaryActive, setIsRotaryActive] = React.useState(false);
+
+  /**
+   * Triggers whenever openModal status updates
+   */
+  useEffect(() => {
+    setOpen(openModal);
+    return () => {
+      setOpen(false);
+    };
+  }, [openModal]);
   /**
    * Initalization to set default value & modal status
    */
   const initSetValue = useCallback(() => {
     setValue(committedSetting >= min ? committedSetting : min);
-    setOpen(openModal);
-  }, [committedSetting, openModal, min]);
+  }, [committedSetting, min]);
 
   useEffect(() => {
     initSetValue();
   }, [initSetValue]);
-
   /**
    * Triggers callback when modal status changes
    */
@@ -129,32 +121,32 @@ export const ValueModal = ({
     }
   });
 
-  /**
-   * Function for handling opening of the model.
-   */
-  const handleOpen = () => {
-    setOpen(true);
-  };
-  /**
-   * Function for handling closing of the model.
-   */
-  const handleClose = () => {
-    setOpen(false);
-  };
-  /**
-   * Function for handling confirmation of the model.
-   */
-  const handleConfirm = () => {
+  useEffect(() => {
     requestCommitSetting(value);
-    setOpen(false);
-  };
+  }, [requestCommitSetting, value]);
+
+  // const handleConfirm = () => {
+  //   requestCommitSetting(value);
+  // };
+
+  /**
+   * Triggers whenever rotaryEncoder value in redux store is updated.
+   */
+  useEffect(() => {
+    // Disable on initial mount since multi step dialog will run everytime when dialog opens
+    if (isInitialMount.current) {
+      isInitialMount.current = false;
+    } else {
+      setIsRotaryActive(true);
+    }
+  }, [rotaryEncoder]);
 
   /**
    * Function for handling label changed for PIP.
    *
-   * @param {string} label - Value label
+   * @param {string} label - desc for label
    *
-   * @returns string
+   * @returns empty string
    */
   function pipClarify(label: string) {
     if (label === 'PIP') return '*not PEEP compensated';
@@ -191,21 +183,9 @@ export const ValueModal = ({
 
   return (
     <Grid container direction="column" alignItems="center" justify="center">
-      <Grid container item xs>
-        {!disableSetNewButton && (
-          <Button
-            onClick={handleOpen}
-            color="primary"
-            variant="contained"
-            className={classes.openButton}
-          >
-            Set New
-          </Button>
-        )}
-      </Grid>
       {rotaryEncoder !== null ? (
         <RotaryEncodeController
-          isActive={true}
+          isActive={isRotaryActive}
           value={value}
           onClick={(num: number) => {
             setValue(num);
@@ -214,15 +194,9 @@ export const ValueModal = ({
           max={max}
         />
       ) : null}
-      <ModalPopup
-        withAction={true}
-        label="Set New"
-        open={open}
-        onClose={handleClose}
-        onConfirm={handleConfirm}
-      >
+      <Grid container direction="column" alignItems="center" justify="center">
         {modalContent}
-      </ModalPopup>
+      </Grid>
     </Grid>
   );
 };
